Create symbol identifiers and support user expressions
We want to let users create default entities by directly manipulating
the linear system of equations. For example let's say a rectangle is
completely constrained by 4 points already in the layout, then adding it
by specifying the size isn't a good option.r

For this we need to syntax for equation symbols that users can use and
our renderer will correctly interpret as such entity. Then we need a way
of parsing and creating these symbols. This is what this commit adds.